Refactor webhook to enable code reuse
