+++ main.go
@@ -28,6 +28,7 @@ import (
 
 	api "github.com/aws/eks-anywhere-packages/api/v1alpha1"
 	"github.com/aws/eks-anywhere-packages/controllers"
+	"github.com/aws/eks-anywhere-packages/pkg/webhook"
 )
 
 var (
@@ -94,7 +95,7 @@ func main() {
 			setupLog.Error(err, "unable to create webhook", "webhook", "ActiveBundleValidator")
 			os.Exit(1)
 		}
-		if err = api.InitPackageValidator(mgr); err != nil {
+		if err = webhook.InitPackageValidator(mgr); err != nil {
 			setupLog.Error(err, "unable to create webhook", "webhook", "Package")
 			os.Exit(1)
 		}
